Refactored manual Registry definitions.
